+++ CHANGELOG.md
@@ -5,7 +5,8 @@ ChangeLog
 2.0.1 (2025-12-26)
 ------------------
 
-- IMPROVEMENT: add option -E for loading .env files
+- REFACTOR: change option -e/--extension to option -p/--plugin
+- IMPROVEMENT: add new option -e for loading .env files
 
 2.0.0 (2025-12-21)
 ------------------